+++ tests/test_paramattr_value.py
@@ -160,6 +160,44 @@ def test_default_values(opened_device, role_type):
         value = getattr(opened_device, attr_def.name)
         assert value == attr_def.default, f'pythonic type of {attr_def.name}'
 
+import inspect
+
+def test_constructor():
+    DEFAULT_VALUE = 4
+
+    class Device(lb.Device):
+        number: int = attr.value.int(default=DEFAULT_VALUE)
+
+    # check Device signature
+    params = inspect.signature(Device).parameters
+
+    assert tuple(params.keys()) == ('resource', 'number'), \
+           'constructor keyword argument names'
+    
+    defaults = [p.default for p in params.values()]
+    assert tuple(defaults) == (lb.Device.resource.default, DEFAULT_VALUE)
+
+
+def test_adjusted_constructor():
+    DEFAULT_VALUE = 4
+
+    class ParentDevice(lb.Device):
+        number: int = attr.value.int(0)
+
+    class Device(ParentDevice):
+        number: int = ParentDevice.number.copy(default=DEFAULT_VALUE)
+
+    # check Device signature
+    params = inspect.signature(Device).parameters
+
+    assert tuple(params.keys()) == ('resource', 'number'), \
+           'constructor keyword argument names'
+    
+    defaults = [p.default for p in params.values()]
+    assert tuple(defaults) == (lb.Device.resource.default, DEFAULT_VALUE)
+
+    d = Device()
+    assert d.number == DEFAULT_VALUE
 
 def test_only(opened_device):
     # low bound