+++ ChangeLog.md
@@ -4,8 +4,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [0.34]
+## [0.35.0]
+### Added
+- `labbench.paramattr.copy` has been to create adjusted copies of other paramattrs.
+  This is meant to replace `labbench.paramattr.adjusted`, which did not update the type hints
+  for default values of Device constructors.
 
+### Deprecated
+- `labbench.paramattr.adjusted`, due to type hinting bugs
+
+
+## [0.34.0]
 ### Changed
 - In order to better align `labbench.paramattr.property` and `labbench.paramattr.method` with each other and the rest of the ecosystem,
   decorator implementation for these attributes now follows [python's built-in syntax for properties](https://docs.python.org/3/library/functions.html#property). For example, to implement getters and setters for a property named `number` or a method named `flag`:
@@ -38,7 +47,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 ### Removed
 - `labbench.paramattr.register_key_argument` was removed in favor of directly decorating Device classes with `labbench.paramattr.method_kwarg`
 
-## [0.33]
+## [0.33.0]
 This is a significant API change.
 
 ### Added